회원과 게시판 관계 추가 및 삭제권한기능

+++ src/main/java/com/train/notice/domain/board/service/BoardService.java
@@ -5,12 +5,15 @@ import java.time.LocalDateTime;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.train.notice.domain.board.entity_dto.Board;
 import com.train.notice.domain.board.entity_dto.BoardDto;
 import com.train.notice.domain.board.repository.BoardRepository;
+import com.train.notice.domain.member.entity_dto.Member;
+import com.train.notice.domain.member.repository.MemberRepository;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
@@ -21,6 +24,7 @@ import lombok.RequiredArgsConstructor;
 public class BoardService {
 
     private final BoardRepository boardRepository;
+    private final MemberRepository memberRepository;
     
     @Transactional(readOnly = true)
     public Page<BoardDto> readBoardList(int page, int size){
@@ -34,9 +38,11 @@ public class BoardService {
     }
 
     @Transactional
-    public BoardDto createBoard(BoardDto boardDto){
-        Board board =Board.create(boardDto.getName(), boardDto.getContent(), LocalDateTime.now(), boardDto.getPoster());
-        return BoardDto.from(boardRepository.save(board));
+    public Long createBoard(BoardDto boardDto, String loginId){
+        Member member = memberRepository.findById(loginId).orElseThrow(()->new IllegalArgumentException("회원이 존재하지 않습니다."));
+
+        Board board =Board.create(boardDto.getName(), boardDto.getContent(), LocalDateTime.now(),member);
+        return boardRepository.save(board).getId();
     }
 
     @Transactional
@@ -48,7 +54,14 @@ public class BoardService {
     }
 
     @Transactional
-    public void deleteBoard(Long id){
+    public void deleteBoard(Long id, String loginId){
+        Board board = boardRepository.findById(id).orElseThrow(()-> new IllegalArgumentException("게시글이 존재하지 않습니다."));
+        Member loginMember = memberRepository.findById(loginId).orElseThrow(()-> new IllegalArgumentException("회원이 존재하지 않습니다"));
+
+        boolean isWriter = board.getWriter().getUsername().equals(loginMember.getUsername());
+        if(!isWriter){
+            throw new AccessDeniedException("삭제 권한이 없습니다.");
+        }
         boardRepository.deleteById(id);
     }
